fix empty raw list to convert

+++ src/main/scala/hu/szigyi/ettl/web/service/ConvertService.scala
@@ -2,7 +2,7 @@ package hu.szigyi.ettl.web.service
 
 import com.typesafe.scalalogging.StrictLogging
 import hu.szigyi.ettl.web.service.ConvertService.{Converter, Converting, Idle, converterRef}
-import hu.szigyi.ettl.web.util.RawToJpg.{convertToJpg, filePathToJpg}
+import hu.szigyi.ettl.web.util.RawToJpg.convertToJpg
 
 import java.util.concurrent.atomic.AtomicReference
 
@@ -26,7 +26,7 @@ class ConvertService(dir: DirectoryService, rawDirectoryPath: => Option[String],
         val raws = files.toList.filter(_.endsWith(rawFileExtension))
 
         jpgs.foreach(jpg => if (!ImageService.containsJpgPath(jpg)) ImageService.addJpgPath(jpg))
-        if (!"jpg".equalsIgnoreCase(rawFileExtension)) {
+        if (raws.nonEmpty && !"jpg".equalsIgnoreCase(rawFileExtension)) {
           if (convertToJpg(raws.last)) dir.deleteFile(raws.last)
         }
       }
